requirements.txt
numpy
opencv-python
nibabel
pydicom
Pillow
boto3
pydantic>=2.0
tqdm
pynrrd>=1.0
